Update Node Sync test and run scripts
